backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
pydantic[email]
